messaging_app/requirements.txt
aiosqlite==0.21.0
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
greenlet==3.2.2
idna==3.10
parameterized==0.9.0
pycodestyle==2.13.0
pycparser==2.22
PyMySQL==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
